requirements.txt
# langchain>=0.1.0
# langchain-community>=0.0.70
# google-generativeai>=0.1.0
# streamlit>=1.25.0
# python-dotenv>=1.0.0
# faiss-cpu>=1.7.4
# langchain-google-genai==2.1.10

# Core
langchain
langchain-community
langchain-google-genai
openai
faiss-cpu
tiktoken

# PDF loading
pypdf
pdfplumber

# UI
streamlit

# Utils
python-dotenv

# Google AI
google-generativeai
